Add Clash.Normalize.Transformations.Letrec
All transformations which are exclusively concerning letrec
expressions now live in a separate module. This is part of work to
make the normalization modules quicker to compile.
