+++ clash-lib/src/Clash/Core/TermInfo.hs
@@ -304,6 +304,11 @@ isPrim :: Term -> Bool
 isPrim (Prim {}) = True
 isPrim _         = False
 
+-- | Is a term a tick?
+isTick :: Term -> Bool
+isTick Tick{} = True
+isTick _ = False
+
 -- | Is a term a cast?
 isCast :: Term -> Bool
 isCast (Cast {}) = True
